fix react example
