+++ examples/react/src/components/atoms/styled/index.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const Content = styled.div`
+  border: solid 1px;
+  padding: 15px;
+  margin: 5px;
+`;
